Canvas Account ID can be much bigger than 2**31-1

+++ integrated_channels/canvas/models.py
@@ -50,7 +50,7 @@ class CanvasEnterpriseCustomerConfiguration(EnterpriseCustomerPluginConfiguratio
         )
     )
 
-    canvas_account_id = models.IntegerField(
+    canvas_account_id = models.BigIntegerField(
         null=True,
         verbose_name="Canvas Account Number",
         help_text="Account number to use during api calls. Called account_id in canvas. "
